委任状アプリのUIを作成し、動的ルーティングを実現
