+++ CommissionApp/src/lib/interfaces.ts
@@ -1,0 +1,10 @@
+export interface FormTemplate {
+  id: string;
+  title: string;
+  description: string;
+  recipientName: string;
+  recipientAddress: string;
+  grantorName: string;
+  grantorAddress: string;
+  createdAt: number;
+}
